[引擎] 提交蒙皮动画相关逻辑

+++ Engine/Source/Runtime/Function/Renderer/Resources/Parsers/AssimpParser.cpp
@@ -179,7 +179,7 @@ LitchiRuntime::Mesh* LitchiRuntime::Parsers::AssimpParser::ProcessMesh(Model* mo
 			vertex.boneWeights[2] = vertexBoneData[i].weights[2];
 
 			for (size_t j = 0; j < NUM_BONES_PER_VERTEX; j++)
-				vertex.boneIndices[j] = vertexBoneData[i].boneIndex[j];
+				vertex.boneIndices[j] = static_cast<int32_t>(vertexBoneData[i].boneIndex[j]);
 		}
 		vertices.push_back(vertex);
